requirements.txt
numpy>=1.20.0
scipy>=1.7.0
torch>=1.10.0
d3rlpy>=2.0.0
gymnasium>=0.29.0
pyyaml>=6.0
